+++ src/main/resources/xsd/ce/v11/ce_sello_digital_cont_elec/SelloDigitalContElec.xsd
@@ -1,0 +1,104 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
+           targetNamespace="www.sat.gob.mx/esquemas/ContabilidadE/1_1/SelloDigitalContElec"
+           elementFormDefault="qualified" attributeFormDefault="unqualified">
+    <xs:element name="SelloDigitalContElec">
+        <xs:annotation>
+            <xs:documentation>
+                Documento requerido para el Sello Digital del SAT que da constancia de recibo del archivo de la
+                contabilidad electrónica.
+            </xs:documentation>
+        </xs:annotation>
+        <xs:complexType>
+            <xs:attribute name="Version" use="required" fixed="1.1">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo requerido para expresar la versión del estándar del Sello Digital.
+                    </xs:documentation>
+                </xs:annotation>
+            </xs:attribute>
+            <xs:attribute name="Folio" use="required">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo requerido para expresar los 22 caracteres del folio asignado por el SAT en la recepción
+                        de los archivos.
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:string">
+                        <xs:whiteSpace value="collapse"/>
+                        <xs:pattern value="[0-9]{3}[1-5][0-9]{2}(01|02|03|04|05|06|07|08|09|10|11|12|13)[0-9]{14}"/>
+                    </xs:restriction>
+                </xs:simpleType>
+            </xs:attribute>
+            <xs:attribute name="RFC" use="required">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo requerido para expresar el RFC del contribuyente que envía los datos
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:string">
+                        <xs:minLength value="12"/>
+                        <xs:maxLength value="13"/>
+                        <xs:whiteSpace value="collapse"/>
+                        <xs:pattern value="[A-ZÑ&amp;]{3,4}[0-9]{2}[0-1][0-9][0-3][0-9][A-Z0-9]?[A-Z0-9]?[0-9A-Z]?"/>
+                    </xs:restriction>
+                </xs:simpleType>
+            </xs:attribute>
+            <xs:attribute name="FechadeSello" use="required">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo requerido para expresar la fecha y hora de la generación del Sello digital del SAT. Se
+                        expresa en la forma aaaa-mm-ddThh:mm:ss, de acuerdo con la especificación ISO 8601
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:dateTime">
+                        <xs:whiteSpace value="collapse"/>
+                    </xs:restriction>
+                </xs:simpleType>
+            </xs:attribute>
+            <xs:attribute name="sello" use="optional">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo opcional para contener el sello digital del archivo de contabilidad electrónica, que
+                        corresponda. El sello deberá ser expresado cómo una cadena de texto en formato Base 64.
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:string">
+                        <xs:whiteSpace value="collapse"/>
+                    </xs:restriction>
+                </xs:simpleType>
+            </xs:attribute>
+            <xs:attribute name="noCertificadoSAT" use="required">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo requerido para expresar el número de serie del certificado del SAT usado para generar
+                        el sello digital.
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:string">
+                        <xs:length value="20"/>
+                        <xs:whiteSpace value="collapse"/>
+                    </xs:restriction>
+                </xs:simpleType>
+            </xs:attribute>
+            <xs:attribute name="selloSAT" use="required">
+                <xs:annotation>
+                    <xs:documentation>
+                        Atributo requerido para contener el sello digital del SAT. El sello deberá ser expresado cómo
+                        una cadena de texto en formato Base 64.
+                    </xs:documentation>
+                </xs:annotation>
+                <xs:simpleType>
+                    <xs:restriction base="xs:string">
+                        <xs:whiteSpace value="collapse"/>
+                    </xs:restriction>
+                </xs:simpleType>
+            </xs:attribute>
+        </xs:complexType>
+    </xs:element>
+</xs:schema>
